+++ src/Krino/Krino.Domain.ConstructiveAdpositionalGrammar/Morphemes/IMorpheme.cs
@@ -1,4 +1,5 @@
 ﻿using Krino.Domain.ConstructiveAdpositionalGrammar.Morphemes.StructuralAttributesArrangement;
+using System.Numerics;
 
 namespace Krino.Domain.ConstructiveAdpositionalGrammar.Morphemes
 {
@@ -15,12 +16,12 @@ namespace Krino.Domain.ConstructiveAdpositionalGrammar.Morphemes
         /// <summary>
         /// Encodes structural attributes the morphem carries.
         /// </summary>
-        ulong Attributes { get; set; }
+        BigInteger Attributes { get; set; }
 
         /// <summary>
         /// Encodes semantic attributes the morpheme carries.
         /// </summary>
-        ulong SememeAttributes { get; set; }
+        BigInteger SememeAttributes { get; set; }
 
         /// <summary>
         /// Returns the grammar character of the morpheme.
